remove int cast on server download to client

+++ src/main/java/pt/fcul/sinf/si003/server/ServerThread.java
@@ -99,7 +99,7 @@ public class ServerThread extends Thread {
     private void downloadFile(File file) throws FileNotFoundException {
         FileInputStream fileInputStream = new FileInputStream(file);
         BufferedInputStream bufferedInputStream = new BufferedInputStream(fileInputStream);
-        cloudSocket.sendStream((int) file.length(), bufferedInputStream);
+        cloudSocket.sendStream(file.length(), bufferedInputStream);
     }
 
     /**
